encoder-decoder/requirements.txt
torch
torchvision
accelerate==1.0.0
arrow==1.3.0
av==13.1.0
beartype==0.19.0
einops==0.8.0
huggingface-hub==0.26.2
moviepy==1.0.3
ninja
numpy==1.24.4
omegaconf==2.3.0
pillow==10.4.0
PyYAML==6.0.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tokenizers==0.20.3
torch-fidelity==0.3.0
torchmetrics==1.5.2
tqdm==4.66.5
transformers==4.46.2

